add Dubstep module and tests
