+++ src/SemanticModels.jl
@@ -81,7 +81,8 @@ end
 
 include("diffeq.jl")
 include("regression.jl")
-include("grfn.jl")
+# include("grfn.jl")
+include("cassette.jl")
 
 """    CombinedModel
